requirements.txt
alembic==1.11.1
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
black==23.7.0
certifi==2023.7.22
cffi==1.15.1
click==8.1.6
cryptography==41.0.3
dataclasses==0.6
dnspython==2.4.1
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.1
graphql-core==3.2.3
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mypy-extensions==1.0.0
orjson==3.9.2
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
psycopg2-binary==2.9.6
pycparser==2.21
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
strawberry-graphql==0.198.0
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
